- couple of typo fixes in comments - compiling tesseract 'legacy engine' for the first time.

+++ src/ccmain/par_control.cpp
@@ -60,7 +60,8 @@ void Tesseract::PrerecAllWordsPar(const std::vector<WordData> &words) {
 #  pragma omp parallel for num_threads(10)
 #endif // _OPENMP
     // NOLINTNEXTLINE(modernize-loop-convert)
-    for (size_t b = 0; b < blobs.size(); ++b) {
+	// Error C3016: index variable in OpenMP 'for' statement must have signed integral type
+    for (int b = 0; b < static_cast<int>(blobs.size()); ++b) {
       *blobs[b].choices =
           blobs[b].tesseract->classify_blob(blobs[b].blob, "par", ScrollView::WHITE, nullptr);
     }
